requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
google-generativeai==0.8.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1